+++ src/app/locales/en/components/share-show/pane-files.js
@@ -3,4 +3,5 @@ export default {
     title: 'Shared files have been deleted',
     content: 'This share points to a file/directory that is no longer there – most likely it has been deleted or moved out of the space.',
   },
+  noSuchItem: 'Cannot load share root',
 };